Refactor : authReducerFunc location change
