+++ src/context/AuthorizerContext.jsx
@@ -1,5 +1,5 @@
 import React ,{useReducer ,createContext ,useContext }from 'react';
-
+import  authReducerFunc from '../reducer/authReducerFunc';
 
 
 
@@ -12,32 +12,7 @@ const AuthorizeProvider = ({children})=>{
         encodedToken :""
     }
 
-
-    const authReducerFunc =( authState , authAction )=>{
-
-        const {type , payload }=authAction;
-        switch(type){
-            case "LOGIN":
-                return {...authState , ...payload.foundUser , loginStatus:true ,encodedToken : payload.encodedToken }
-            case "SIGNUP":
-                return {...authState , ...payload.createdUser , loginStatus:true ,encodedToken : payload.encodedToken }
-            case "LOGOUT":
-                return {
-                    loginStatus : false ,
-                    encodedToken :""
-                } 
-            default:
-                return userState;
-
-        }
-
-
-    }
-
-
     const [ authState , authDispatch ] = useReducer( authReducerFunc , initialAuthState);
-
-    console.log("mai hu state", authState);
 
     return<>
         <AuthorizerContext.Provider value={{ authState , authDispatch }} >
